+++ src/yaml.rs
@@ -18,9 +18,8 @@ use yaml_rust::YamlLoader;
 pub struct ParseYaml {
     pub files: Vec<(String, yaml_rust::Yaml)>,
     pub rulecounter: HashMap<String, u128>,
-    pub exclude_rule_count: u128,
-    pub noisy_rule_count: u128,
-    pub deprecate_rule_count: u128,
+    pub rule_load_cnt: HashMap<String, u128>,
+    pub rule_status_cnt: HashMap<String, u128>,
     pub errorrule_count: u128,
 }
 
@@ -35,9 +34,11 @@ impl ParseYaml {
         ParseYaml {
             files: Vec::new(),
             rulecounter: HashMap::new(),
-            exclude_rule_count: 0,
-            noisy_rule_count: 0,
-            deprecate_rule_count: 0,
+            rule_load_cnt: HashMap::from([
+                ("excluded".to_string(), 0_u128),
+                ("noisy".to_string(), 0_u128),
+            ]),
+            rule_status_cnt: HashMap::from([("deprecated".to_string(), 0_u128)]),
             errorrule_count: 0,
         }
     }
@@ -221,19 +222,18 @@ impl ParseYaml {
                 //除外されたルールは無視する
                 let rule_id = &yaml_doc["id"].as_str();
                 if rule_id.is_some() {
-                    match exclude_ids
+                    if let Some(v) = exclude_ids
                         .no_use_rule
-                        .get(&rule_id.unwrap_or("").to_string())
+                        .get(&rule_id.unwrap_or(&String::default()).to_string())
                     {
-                        None => (),
-                        Some(v) => {
-                            if v.contains("exclude_rule") {
-                                self.exclude_rule_count += 1;
-                            } else {
-                                self.noisy_rule_count += 1;
-                            }
-                            return Option::None;
-                        }
+                        let entry_key = if v.contains("exclude_rule") {
+                            "excluded"
+                        } else {
+                            "noisy"
+                        };
+                        let entry = self.rule_load_cnt.entry(entry_key.to_string()).or_insert(0);
+                        *entry += 1;
+                        return Option::None;
                     }
                 }
 
@@ -244,6 +244,17 @@ impl ParseYaml {
                         .unwrap_or(&0)
                         + 1,
                 );
+
+                let status_cnt = self
+                    .rule_status_cnt
+                    .entry(
+                        yaml_doc["status"]
+                            .as_str()
+                            .unwrap_or("undefined")
+                            .to_string(),
+                    )
+                    .or_insert(0);
+                *status_cnt += 1;
 
                 if configs::CONFIG.read().unwrap().args.is_present("verbose") {
                     println!("Loaded yml file path: {}", filepath);
@@ -269,7 +280,11 @@ impl ParseYaml {
                 {
                     let rule_status = &yaml_doc["status"].as_str().unwrap_or_default();
                     if *rule_status == "deprecated" {
-                        self.deprecate_rule_count += 1;
+                        let entry = self
+                            .rule_status_cnt
+                            .entry(rule_status.to_string())
+                            .or_insert(0);
+                        *entry += 1;
                         return Option::None;
                     }
                 }
@@ -400,7 +415,7 @@ mod tests {
         let mut yaml = yaml::ParseYaml::new();
         let path = Path::new("test_files/rules/yaml");
         yaml.read_dir(path, "", &filter::exclude_ids()).unwrap();
-        assert_eq!(yaml.exclude_rule_count, 5);
+        assert_eq!(yaml.rule_load_cnt.get("excluded").unwrap().to_owned(), 5);
     }
     #[test]
     fn test_all_noisy_rules_file() {
@@ -409,7 +424,7 @@ mod tests {
         let mut yaml = yaml::ParseYaml::new();
         let path = Path::new("test_files/rules/yaml");
         yaml.read_dir(path, "", &filter::exclude_ids()).unwrap();
-        assert_eq!(yaml.noisy_rule_count, 5);
+        assert_eq!(yaml.rule_load_cnt.get("noisy").unwrap().to_owned(), 5);
     }
     #[test]
     fn test_none_exclude_rules_file() {
@@ -419,7 +434,7 @@ mod tests {
         let path = Path::new("test_files/rules/yaml");
         let exclude_ids = RuleExclude::default();
         yaml.read_dir(path, "", &exclude_ids).unwrap();
-        assert_eq!(yaml.exclude_rule_count, 0);
+        assert_eq!(yaml.rule_load_cnt.get("excluded").unwrap().to_owned(), 0);
     }
     #[test]
     fn test_exclude_deprecated_rules_file() {
@@ -427,6 +442,9 @@ mod tests {
         let path = Path::new("test_files/rules/deprecated");
         let exclude_ids = RuleExclude::default();
         yaml.read_dir(path, "", &exclude_ids).unwrap();
-        assert_eq!(yaml.deprecate_rule_count, 1);
+        assert_eq!(
+            yaml.rule_status_cnt.get("deprecated").unwrap().to_owned(),
+            2
+        );
     }
 }