+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## v1.3.1 [2022/xx/xx]
 
+**New Features:**
+
+- You can now specify specific fields when there are multiple fields with the same name (Ex: `Data`). In the `details` line in a rule, specify a placeholder like `%Data[1]%` to display the first `Data` field. (#487) (@hitenkoku)
+
 **Enhancements:**
 
 - Strip debug symbols by default for smaller Linux and macOS binaries. (#568) (@YamatoSecurity)
+
+**Bug Fixes:**
+
+- fixed bug that RecordID and RecordInformation column is showed when options is not enabled. (#577) (@hitenkoku)
 
 ## v1.3.0 [2022/06/06]