+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - Strip debug symbols by default for smaller Linux and macOS binaries. (#568) (@YamatoSecurity)
 - Added new output time format options. (`--US-time`, `--US-military-time`, `--European-time`) (#574) (@hitenkoku)
 - Changed output time format when `--rfc-3339` option is enabled. (#574) (@hitenkoku)
-- Changed the `-R / --display-record-id` option to `-R / --hide-record-id`: Default display the event record ID. RecordID is not display with `--hide-record-id` option. (#579) (@hitenkoku)
+- Changed the `-R / --display-record-id` option to `-R / --hide-record-id` and now by default the event record ID is displayed. You can hide the record ID with `-R / --hide-record-id`. (#579) (@hitenkoku)
 
 **Bug Fixes:**