+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **New Features:**
 
 - You can now specify specific fields when there are multiple fields with the same name (Ex: `Data`). In the `details` line in a rule, specify a placeholder like `%Data[1]%` to display the first `Data` field. (#487) (@hitenkoku)
+- Added loaded rules status summary. (#583) (@hitenkoku)
 
 **Enhancements:**
 
@@ -12,6 +13,7 @@
 - Added new output time format options. (`--US-time`, `--US-military-time`, `--European-time`) (#574) (@hitenkoku)
 - Changed output time format when `--rfc-3339` option is enabled. (#574) (@hitenkoku)
 - Changed the `-R / --display-record-id` option to `-R / --hide-record-id` and now by default the event record ID is displayed. You can hide the record ID with `-R / --hide-record-id`. (#579) (@hitenkoku)
+- Added rule loading message. (#583) (@hitenkoku)
 
 **Bug Fixes:**